Lift some stuff into input.c
