+++ input.c
@@ -1,0 +1,156 @@
+#include <ncurses.h>
+
+#include "game.h"
+#include "input.h"
+
+void print_location_elements(level *lvl, mobile *mob) {
+    char *message = malloc(sizeof(char)*MESSAGE_LENGTH);
+    constituents *chem = lvl->chemistry[mob->x][mob->y];
+    snprintf(message, MESSAGE_LENGTH, "wood: %d air: %d fire: %d", chem->elements[wood], chem->elements[air], chem->elements[fire]);
+    print_message(message);
+    free((void*)message);
+}
+
+void toggle_door(level *lvl, mobile *mob) {
+    int x = mob->x;
+    int y = mob->y;
+    switch (getch()) {
+        case KEY_UP:
+            y -= 1;
+            break;
+        case KEY_DOWN:
+            y += 1;
+            break;
+        case KEY_RIGHT:
+            x += 1;
+            break;
+        case KEY_LEFT:
+            x -= 1;
+            break;
+        default:
+            return;
+    }
+    if (lvl->tiles[x][y] == OPEN_DOOR) {
+        lvl->tiles[x][y] = CLOSED_DOOR;
+    } else if (lvl->tiles[x][y] == CLOSED_DOOR) {
+        lvl->tiles[x][y] = OPEN_DOOR;
+    }
+}
+
+void drop_item(level *lvl, mobile *mob) {
+    item *item = pop_inventory(mob);
+    if (item != NULL) {
+        level_push_item(lvl, item, mob->x, mob->y);
+        if (mob == lvl->player) {
+            char msg[MESSAGE_LENGTH];
+            snprintf(msg, MESSAGE_LENGTH, "You drop the %s", item->name);
+            print_message(msg);
+        }
+    }
+}
+
+void pickup_item(level *lvl, mobile *mob) {
+    item *item = level_pop_item(lvl, mob->x, mob->y);
+    if (item != NULL) {
+        push_inventory(mob, item);
+        if (mob == lvl->player) {
+            char msg[MESSAGE_LENGTH];
+            snprintf(msg, MESSAGE_LENGTH, "You pick up the %s", item->name);
+            print_message(msg);
+        }
+    }
+}
+
+void smash(level *lvl, mobile *mob) {
+    item *potion = ((item*)mob)->contents->item;
+    add_constituents(lvl->chemistry[mob->x][mob->y], potion->chemistry);
+    inventory_item *inv = ((item*)mob)->contents;
+    ((item*)mob)->contents = inv->next;
+    free((void*)inv);
+    destroy_item(potion);
+}
+
+
+int get_input(level *lvl) {
+    //TODO Don't advance turn on any key press.
+    //Have explicit "wait" key (e.g. <SPACE>)
+    char *inventory;
+    char *message = malloc(sizeof(char)*MESSAGE_LENGTH);
+    int input = getch();
+    int return_code = 0;
+
+    switch (input) {
+        case KEY_UP:
+            lvl->keyboard_y = -1;
+            break;
+        case KEY_DOWN:
+            lvl->keyboard_y = 1;
+            break;
+        case KEY_RIGHT:
+            lvl->keyboard_x = 1;
+            break;
+        case KEY_LEFT:
+            lvl->keyboard_x = -1;
+            break;
+        case 'H':
+            snprintf(message, MESSAGE_LENGTH, "HELP: (i)nv (d)rop (.)pickup (r)otate inv (q)uaff (v)smash (e)xamine (s)tatus (t)ile (Q)uit");
+            print_message(message);
+            break;
+        case 'i':
+            inventory = inventory_string(lvl->player, MESSAGE_LENGTH);
+            snprintf(message, MESSAGE_LENGTH, "Your inventory contains: %s", inventory);
+            print_message(message);
+            free((void*)inventory);
+            break;
+        case 'd':
+            drop_item(lvl, lvl->player);
+            break;
+        case '.':
+            pickup_item(lvl, lvl->player);
+            break;
+        case 'r':
+            rotate_inventory(lvl->player);
+            inventory = inventory_string(lvl->player, MESSAGE_LENGTH);
+            snprintf(message, MESSAGE_LENGTH, "Your inventory contains: %s", inventory);
+            print_message(message);
+            free((void*)inventory);
+            break;
+        case 'q':
+            if(quaff(lvl->player)) {
+                print_message("You drink a potion.");
+            } else {
+                print_message("That isn't a potion.");
+            }
+            break;
+        case 'v':
+            if (((item*)lvl->player)->contents != NULL && ((item*)lvl->player)->contents->item->type == Potion) {
+                smash(lvl, lvl->player);
+                print_message("You smash the potion on the floor.");
+            } else print_message("That isn't a potion.");
+            break;
+        case 'e':
+            if (((item*)lvl->player)->contents != NULL) {
+                snprintf(message, MESSAGE_LENGTH, "wood: %d fire: %d ash: %d",
+                        ((item*)lvl->player)->contents->item->chemistry->elements[wood],
+                        ((item*)lvl->player)->contents->item->chemistry->elements[fire],
+                        ((item*)lvl->player)->contents->item->chemistry->elements[ash]
+                        );
+                print_message(message);
+            }
+            break;
+        case 's':
+            snprintf(message, MESSAGE_LENGTH, "You have %d hit points. venom: %d banz: %d life: %d", ((item*)lvl->player)->health, ((item*)lvl->player)->chemistry->elements[venom], ((item*)lvl->player)->chemistry->elements[banz], ((item*)lvl->player)->chemistry->elements[life]);
+            print_message(message);
+            break;
+        case 't':
+            print_location_elements(lvl, lvl->player);
+            break;
+        case 'o':
+            toggle_door(lvl, lvl->player);
+            break;
+        case 'Q':
+            return_code = -1;
+    }
+    free((void*)message);
+    return return_code;
+}
