+++ input.h
@@ -1,0 +1,7 @@
+#ifndef INPUT_H
+#define INPUT_H
+
+#include "level/level.h"
+
+int get_input(level *lvl);
+#endif
